+++ mimesis/providers/text.py
@@ -1,6 +1,8 @@
 """Provides data related to text."""
+import typing as t
 
 from mimesis.data import SAFE_COLORS
+from mimesis.enums import EmojyCategory
 from mimesis.providers.base import BaseDataProvider
 
 __all__ = ["Text"]
@@ -9,12 +11,17 @@ __all__ = ["Text"]
 class Text(BaseDataProvider):
     """Class for generating text data."""
 
+    def __init__(self, *args: t.Any, **kwargs: t.Any) -> None:
+        """Initialize attributes."""
+        super().__init__(*args, **kwargs)
+        self._emojis = self._read_global_file("emojis.json")
+
     class Meta:
         name = "text"
         datafile = f"{name}.json"
 
     def alphabet(self, lower_case: bool = False) -> list[str]:
-        """Get an alphabet for current locale.
+        """Returns an alphabet for current locale.
 
         :param lower_case: Return alphabet in lower case.
         :return: Alphabet.
@@ -25,7 +32,7 @@ class Text(BaseDataProvider):
         return alpha
 
     def level(self) -> str:
-        """Generate a random level of danger or something else.
+        """Generates a word that indicates a level of something.
 
         :return: Level.
 
@@ -36,7 +43,7 @@ class Text(BaseDataProvider):
         return self.random.choice(levels)
 
     def text(self, quantity: int = 5) -> str:
-        """Generate the text.
+        """Generates the text.
 
         :param quantity: Quantity of sentences.
         :return: Text.
@@ -45,21 +52,21 @@ class Text(BaseDataProvider):
         return " ".join(self.random.choices(text, k=quantity))
 
     def sentence(self) -> str:
-        """Get a random sentence from text.
+        """Generates a random sentence from the text.
 
         :return: Sentence.
         """
         return self.text(quantity=1)
 
     def title(self) -> str:
-        """Get a random title.
+        """Generates a random title.
 
         :return: The title.
         """
         return self.text(quantity=1)
 
     def words(self, quantity: int = 5) -> list[str]:
-        """Generate a list of random words.
+        """Generates a list of random words.
 
         :param quantity: Quantity of words. Default is 5.
         :return: Word list.
@@ -71,7 +78,7 @@ class Text(BaseDataProvider):
         return self.random.choices(words, k=quantity)
 
     def word(self) -> str:
-        """Get a random word.
+        """Generates a random word.
 
         :return: Single word.
 
@@ -81,9 +88,9 @@ class Text(BaseDataProvider):
         return self.words(quantity=1)[0]
 
     def quote(self) -> str:
-        """Get a random quote.
+        """Generates a random quote.
 
-        :return: Quote from movie.
+        :return: Random quote.
 
         :Example:
             "Bond... James Bond."
@@ -92,7 +99,7 @@ class Text(BaseDataProvider):
         return self.random.choice(quotes)
 
     def color(self) -> str:
-        """Get a random name of color.
+        """Generates a random color name.
 
         :return: Color name.
 
@@ -104,7 +111,7 @@ class Text(BaseDataProvider):
 
     @staticmethod
     def _hex_to_rgb(color: str) -> tuple[int, ...]:
-        """Convert hex color to RGB format.
+        """Converts hex color to RGB format.
 
         :param color: Hex color.
         :return: RGB tuple.
@@ -113,7 +120,7 @@ class Text(BaseDataProvider):
         return tuple(int(color[i : i + 2], 16) for i in (0, 2, 4))
 
     def hex_color(self, safe: bool = False) -> str:
-        """Generate a random hex color.
+        """Generates a random HEX color.
 
         :param safe: Get safe Flat UI hex color.
         :return: Hex color code.
@@ -127,7 +134,7 @@ class Text(BaseDataProvider):
         return f"#{self.random.randint(0x000000, 0xFFFFFF):06x}"
 
     def rgb_color(self, safe: bool = False) -> tuple[int, ...]:
-        """Generate a random rgb color tuple.
+        """Generates a random RGB color tuple.
 
         :param safe: Get safe RGB tuple.
         :return: RGB tuple.
@@ -139,7 +146,7 @@ class Text(BaseDataProvider):
         return self._hex_to_rgb(color)
 
     def answer(self) -> str:
-        """Get a random answer in the current language.
+        """Generates a random answer in the current language.
 
         :return: An answer.
 
@@ -148,3 +155,25 @@ class Text(BaseDataProvider):
         """
         answers: list[str] = self.extract(["answers"])
         return self.random.choice(answers)
+
+    def emoji(self, category: EmojyCategory | None = EmojyCategory.DEFAULT) -> str:
+        """Generates a random emoji from the specified category.
+
+        Generates a random emoji from the specified category.
+        If the category is not specified, a random emoji
+        from any category will be returned.
+
+        :param category: :class:`~mimesis.enums.EmojyCategory`.
+        :raises NonEnumerableError: When category is not supported.
+        :return: Emoji code.
+        :example:
+            😟
+        """
+        category = self.validate_enum(category, EmojyCategory)
+        symbol = self.random.choice(self._emojis[category])
+
+        # Some emoji consist of multiple Unicode characters.
+        if isinstance(symbol, list):
+            return "".join([chr(int(s, 16)) for s in symbol])
+
+        return chr(int(symbol, 16))